+++ modules/table/src/integrationTest/java/org/apache/ignite/distributed/ItTxDistributedTestSingleNodeNoCleanupMessage.java
@@ -73,7 +73,7 @@ import org.junit.jupiter.api.TestInfo;
 /**
  * Test to Simulate missing cleanup action.
  */
-public class ItTxDistributedTestSingleNodeNoCleanupMessage extends ItTxDistributedTestSingleNode {
+public class ItTxDistributedTestSingleNodeNoCleanupMessage extends ItTxAbstractDistributedTest {
     /** A list of background cleanup futures. */
     private final List<CompletableFuture<?>> cleanupFutures = new CopyOnWriteArrayList<>();
